Fixed a bug in the GetPDFText Command
The GetPDFText command was non-functional for online PDF's prior to the implementation of the gallery manager. This seems to be because webclient connections fail when the global ServicePointHandler is set to its default protocol of ssl3. The gallery manager constructor sets this value to Tls12, which unknowingly fixed this issue in the GetPDFText command after its implementation. I have added this code to the command's constructor so that it is not reliant on the unrelated gallery manager.